Implementando outras exceções customizadas

+++ src/main/java/com/campospadilhaa/dscommerce/services/ProductService.java
@@ -4,15 +4,20 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.campospadilhaa.dscommerce.dto.ProductDTO;
 import com.campospadilhaa.dscommerce.entities.Product;
 import com.campospadilhaa.dscommerce.repositories.ProductRepository;
+import com.campospadilhaa.dscommerce.services.exceptions.DatabaseException;
 import com.campospadilhaa.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -76,18 +81,32 @@ public class ProductService {
 	@Transactional
 	public ProductDTO update(Long id, ProductDTO productDTO) {
 
-		Product product = productRepository.getReferenceById(id); // método 'getReferenceById' não vai no banco de dados
+		try {
+			
+			Product product = productRepository.getReferenceById(id); // método 'getReferenceById' não vai no banco de dados
 
-		copyDtoToEntity(productDTO, product);
+			copyDtoToEntity(productDTO, product);
 
-		product = productRepository.save(product);
+			product = productRepository.save(product);
 
-		return new ProductDTO(product);
+			return new ProductDTO(product);
+
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Produto não encontrado para atualização");
+		}
 	}
 
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
 
-		productRepository.deleteById(id);
+		if (!productRepository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+
+		try {
+			productRepository.deleteById(id);    		
+		}catch (DataIntegrityViolationException e) {
+	        throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
 }
